+++ R/se.R
@@ -17,16 +17,22 @@ utils::globalVariables(c("strap", "models", "estimate"))
 #' @param nsim Numeric, the number of simulations for calculating the
 #'          standard error for intraclass correlation coefficients, as
 #'          obtained by the \code{\link{icc}}-function.
+#' @param type Type of standard errors for generalized linear mixed models.
+#'          \code{type = "fe"} returns the standard errors for fixed effects,
+#'          based on the delta-method-approximation. \code{type = "se"} returns
+#'          the standard errors for joint random and fixed effects. See 'Details'.
 #'
 #' @return The standard error of \code{x}.
 #'
 #' @note Computation of standard errors for coefficients of mixed models
 #'         is based \href{http://stackoverflow.com/questions/26198958/extracting-coefficients-and-their-standard-error-from-lme}{on this code}.
 #'         \cr \cr
-#'         Standard errors for generalized linear (mixed) models are
-#'         approximations based on the delta method (Oehlert 1992).
+#'         Standard errors for generalized linear (mixed) models, if
+#'            \code{type = "re"}, are approximations based on the delta
+#'            method (Oehlert 1992).
 #'
-#' @details For linear mixed models, this function computes the standard errors
+#' @details For linear mixed models, and generalized linear mixed models with
+#'            \code{type = "re"}, this function computes the standard errors
 #'            for joint (sums of) random and fixed effects coefficients (unlike
 #'            \code{\link[arm]{se.coef}}, which returns the standard error
 #'            for fixed and random effects separately). Hence, \code{se()}
@@ -35,8 +41,11 @@ utils::globalVariables(c("strap", "models", "estimate"))
 #'            For generalized linear models or generalized linear mixed models,
 #'            approximated standard errors, using the delta method for transformed
 #'            regression parameters are returned (Oehlert 1992). For generalized
-#'            linear mixed models, the standard errors refer to the fixed effects
-#'            only.
+#'            linear mixed models, by default, the standard errors refer to the
+#'            fixed effects only. Use \code{type = "re"} to compute standard errors
+#'            for joint random and fixed effects coefficients. However, this
+#'            computation \emph{is not} based on the delta method, so standard
+#'            errors from \code{type = "re"} are on the logit-scale.
 #'            \cr \cr
 #'            The standard error for the \code{\link{icc}} is based on bootstrapping,
 #'            thus, the \code{nsim}-argument is required. See 'Examples'.
@@ -109,7 +118,10 @@ utils::globalVariables(c("strap", "models", "estimate"))
 #' @importFrom broom tidy
 #' @importFrom dplyr mutate select_
 #' @export
-se <- function(x, nsim = 100) {
+se <- function(x, nsim = 100, type = c("fe", "re")) {
+  # match arguments
+  type <- match.arg(type)
+
   if (inherits(x, c("lmerMod", "nlmerMod", "merModLmerTest"))) {
     # return standard error for (linear) mixed models
     return(std_merMod(x))
@@ -117,19 +129,26 @@ se <- function(x, nsim = 100) {
     # we have a ICC object, so do bootstrapping and compute SE for ICC
     return(std_e_icc(x, nsim))
   } else if (inherits(x, c("glm", "glmerMod"))) {
-    # for glm, we want to exponentiate coefficients to get odds ratios, however
-    # 'exponentiate'-argument currently not works for lme4-tidiers
-    # so we need to do this manually for glmer's
-    tm <- broom::tidy(x, effects = "fixed")
-    tm$estimate <- exp(tm$estimate)
-    return(
-      tm %>%
-        # vcov for merMod returns a dpoMatrix-object, so we need
-        # to coerce to regular matrix here.
-        dplyr::mutate(or.se = sqrt(estimate ^ 2 * diag(as.matrix(stats::vcov(x))))) %>%
-        dplyr::select_("term", "estimate", "or.se") %>%
-        sjmisc::var_rename(or.se = "std.error")
-    )
+    # check type of se
+    if (type == "fe") {
+      # for glm, we want to exponentiate coefficients to get odds ratios, however
+      # 'exponentiate'-argument currently not works for lme4-tidiers
+      # so we need to do this manually for glmer's
+      tm <- broom::tidy(x, effects = "fixed")
+      tm$estimate <- exp(tm$estimate)
+      return(
+        tm %>%
+          # vcov for merMod returns a dpoMatrix-object, so we need
+          # to coerce to regular matrix here.
+          dplyr::mutate(or.se = sqrt(estimate ^ 2 * diag(as.matrix(stats::vcov(x))))) %>%
+          dplyr::select_("term", "estimate", "or.se") %>%
+          sjmisc::var_rename(or.se = "std.error")
+      )
+    } else {
+      # return standard error for mixed models,
+      # joint random and fixed effects
+      return(std_merMod(x))
+    }
   } else if (inherits(x, "lm")) {
     # for convenience reasons, also return se for simple linear models
     return(x %>%
@@ -167,7 +186,7 @@ se <- function(x, nsim = 100) {
 std_e_helper <- function(x) sqrt(var(x, na.rm = TRUE) / length(stats::na.omit(x)))
 
 
-#' @importFrom stats coef setNames
+#' @importFrom stats coef setNames vcov
 #' @importFrom lme4 ranef
 std_merMod <- function(fit) {
   se.merMod <- list()
@@ -179,7 +198,7 @@ std_merMod <- function(fit) {
   inames <- names(cc)
 
   # variances of fixed effects
-  fixed.vars <- diag(as.matrix(lme4::vcov.merMod(fit)))
+  fixed.vars <- diag(as.matrix(stats::vcov(fit)))
 
   # extract variances of conditional modes
   r1 <- lme4::ranef(fit, condVar = TRUE)